Improved search
Now, if the query starts with #, it is interpreted
to mean a member id. Like, #55 redirects directly
to member 55 editing.

Secondly, the search query is split from space
so that searching for "alina suominen" will match
only contacts with both terms "alina" and
"suominen".

Last improvement is that if there is only one
result, the search form redirects directly to the
editing page.