Cleaned up default settings
Moved lots of stuff from local_settings.py.template
to devel_settings.py and put dummy values in place
so that it is very quick to get started with
sikteeri. No need to think what is the IBAN format.

Changed default DEBUG to False. Undefined default
database; it must be specified in local_settings.