Make due_date monotonically increasing
For every bill, due date is derived from the time
the bill is sent.

SingleMemberBillingModelsTest.test_billing_cycle_last_bill
assumed that the reminder it just sent is the last
bill by due date (and we use last_bill elsewhere
too). This was not correct as they both had identical
due dates and database was free to choose the wrong
bill.

It probably only matters in case of unit tests
where both the bill and its reminder are forcibly
generated on the same day.

An alternate fix for last_bill could be to sort bills
by both due date and reminder count, but I think
that would make it unnecessarily more complex as
we can hide the information in the seconds field.