PDF bill: List payments & refactor VAT logic
For some reason VAT calculation was different
implementation from that in models.Bill.
Copied models implementation here, and also
refactored the naming to be consistent with
what is used in models.

I also changed bills to lineitems since that is
what it really is; it is the list of lines in
PDF.

This fixes kapsiry/sikteeri#117.